requirements.txt
numpy>=1.24.0
sounddevice>=0.4.0
python-vlc>=3.0.18
pyserial>=3.5
pillow>=9.0.0
librosa>=0.10.0
mypy>=1.8.0
screeninfo>=0.8.1
pytest>=8.0.0
